+++ snowchains_core/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## [Unreleased]
+## [0.11.0] - 2021-02-26Z
 
 ### Changed